avoid circular dependencies and also add top level file to require
