+++ library/cwm/src/lib/cwm.rb
@@ -1,0 +1,1 @@
+require "cwm/widget"
